Get the handle of ConnMan service based on the profile name
For getting the correct profile name, need to combine essid,
wlan_mode and sec_mode fields of struct net_profile_info_t
to a group name based on the provided algorithm of ConnMan,
and then find out the correct path of ConnMan service from
ConnMan service set via this group name, in CAPI the path of
ConnMan service has been renamed profile_name(profile name).

The profile name(the path of ConnMan service) can be as a
unique identification, so the program can get the handle of
ConnMan service based on the profile name.

Change-Id: Id6ddc1985d6ee0398312fcf0409dd721efb2f46a
Signed-off-by: Chengyi Zhao <chengyi1.zhao@archermind.com>